requirement_conda.0901.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backcall=0.1.0=py37_0
blas=1.0=mkl
ca-certificates=2019.5.15=1
certifi=2019.6.16=py37_1
cffi=1.12.1=py37h2e261b9_0
cudatoolkit=9.0=h13b8566_0
decorator=4.3.2=pypi_0
freetype=2.9.1=h8a8886c_1
ignite=0.2.0=py37_0
intel-openmp=2019.1=144
ipykernel=5.1.0=py37h39e3cac_0
ipython=7.4.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
isodate=0.6.0=pypi_0
jedi=0.13.3=py37_0
jpeg=9b=h024ee3a_2
jupyter_client=5.2.4=py37_0
jupyter_core=4.4.0=py37_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.36=hbc83047_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
mkl=2019.1=144
mkl_fft=1.0.10=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
ncurses=6.1=he6710b0_1
networkx=2.2=pypi_0
ninja=1.8.2=py37h6bb024c_1
numpy=1.16.2=py37h7e9f1db_0
numpy-base=1.16.2=py37hde5b4d6_0
olefile=0.46=py37_0
openssl=1.1.1c=h7b6447c_1
pandas=0.24.1=pypi_0
parso=0.3.4=py37_0
pexpect=4.6.0=py37_0
pickleshare=0.7.5=py37_0
pillow=5.4.1=py37h34e0f95_0
pip=19.0.3=py37_0
plyfile=0.7=pypi_0
prompt_toolkit=2.0.9=py37_0
ptyprocess=0.6.0=py37_0
pycparser=2.19=py37_0
pygments=2.3.1=py37_0
pyparsing=2.3.1=pypi_0
python=3.7.2=h0371630_0
python-dateutil=2.8.0=py37_0
pytorch=1.0.1=py3.7_cuda9.0.176_cudnn7.4.2_2
pytz=2018.9=pypi_0
pyzmq=18.0.0=py37he6710b0_0
rdflib=4.2.2=pypi_0
readline=7.0=h7b6447c_5
scikit-learn=0.20.3=py37hd81dba3_0
scipy=1.2.1=pypi_0
setuptools=40.8.0=py37_0
six=1.12.0=py37_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torch-cluster=1.2.4=pypi_0
torch-geometric=1.0.3=pypi_0
torch-scatter=1.1.2=pypi_0
torch-sparse=0.2.4=pypi_0
torch-spline-conv=1.0.6=pypi_0
torchvision=0.2.2=py_3
tornado=6.0.2=py37h7b6447c_0
tqdm=4.31.1=py_0
traitlets=4.3.2=py37_0
wcwidth=0.1.7=py37_0
wheel=0.33.1=py37_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
